--- NODEJS-16-COMPATIBILITY.md
@@ -1,61 +1,0 @@
-# Node.js 16 Compatibility Fixes
-
-This document outlines all the fixes implemented to ensure the project builds and runs correctly on Node.js 16.
-
-## Issues Addressed
-
-### 1. ReadableStream Not Defined
-**Problem**: Node.js 16 doesn't have the ReadableStream Web API built-in.
-**Solution**: Added `web-streams-polyfill` dependency and polyfill injection in multiple scripts.
-
-### 2. OpenSSL Legacy Provider Errors
-**Problem**: Node.js 16 doesn't support the `--openssl-legacy-provider` flag.
-**Solution**: Modified build script to detect Node.js version and only use legacy provider for Node.js 17+.
-
-### 3. Cheerio Compatibility
-**Problem**: `cheerio` package might have compatibility issues with Node.js 16.
-**Solution**: Added fallback HTML manipulation using string replacement when cheerio fails.
-
-## Files Modified
-
-### Core Build Scripts
-- `scripts/build-with-legacy-support.js` - Smart build script with Node.js version detection
-- `scripts/node16-polyfills.js` - Dedicated polyfill loader for Node.js 16
-- `scripts/prepare-chrome-build.js` - Added ReadableStream polyfill and cheerio fallback
-
-### Package Dependencies
-- `package.json` - Added `web-streams-polyfill` as devDependency
-
-### CI/CD Integration
-- `.github/workflows/ci.yml` - Build matrix tests Node.js 16, 18, 20, 23
-- `.github/workflows/code-quality.yml` - Uses build-with-legacy-support.js
-
-## Implementation Details
-
-### Polyfill Strategy
-1. Check Node.js version using `parseInt(process.version.slice(1))`
-2. If Node.js < 18, attempt to load web-streams-polyfill
-3. Gracefully continue if polyfill not available
-4. Apply polyfills globally when needed
-
-### Build Strategy
-1. Try standard `npm run build` first
-2. If Node.js 17+, try with `--openssl-legacy-provider`
-3. If Node.js 20+, try with additional memory settings
-4. Apply polyfills as fallback for Node.js 16
-5. Log which strategy succeeded for debugging
-
-### HTML Manipulation Fallback
-1. Try using cheerio for robust DOM manipulation
-2. Fall back to string replacement if cheerio fails
-3. Maintain functionality across Node.js versions
-
-## Testing
-- All scripts tested locally on multiple Node.js versions
-- CI matrix validates builds on Node.js 16, 18, 20, 23
-- Both Chrome and Firefox builds supported
-
-## Future Considerations
-- Monitor for additional Node.js 16 compatibility issues
-- Consider upgrading minimum Node.js version in future releases
-- Keep polyfills updated as needed
