--- SECURITY-UPGRADES.md
@@ -1,120 +1,0 @@
-# Security Upgrades and Fixes
-
-This document outlines the security improvements made to the CI/CD pipeline and codebase.
-
-## Semgrep Upgrade
-
-### Problem
-- **Deprecated Action**: Using `returntocorp/semgrep-action@v1` (deprecated)
-- **Old Version**: Running Semgrep 1.36.0 (unsupported, requires 1.76.0+)
-- **No SARIF Output**: Missing integration with GitHub Code Scanning
-
-### Solution
-- **Modern CLI**: Replaced deprecated action with native Semgrep CLI
-- **Latest Version**: Direct installation via pip ensures latest version
-- **SARIF Integration**: Generates and uploads SARIF to GitHub Code Scanning
-- **Enhanced Coverage**: Maintains same ruleset (security-audit, secrets, owasp-top-ten)
-
-### New Implementation
-```yaml
-- name: Install Semgrep 🔍
-  run: |
-    python -m pip install --upgrade pip
-    pip install semgrep
-
-- name: Run Semgrep Scan 🔍
-  run: |
-    semgrep ci --config=p/security-audit --config=p/secrets --config=p/owasp-top-ten --sarif --output=semgrep-results.sarif
-
-- name: Upload Semgrep SARIF 📊
-  uses: github/codeql-action/upload-sarif@v3
-  with:
-    sarif_file: semgrep-results.sarif
-    category: semgrep
-```
-
-## XSS Vulnerability Fix
-
-### Problem
-- **Security Finding**: Potential XSS in `prepare-chrome-build.js`
-- **Root Cause**: Using regex match result directly in script tag context
-- **Risk**: Could inject arbitrary content if HTML is malformed
-
-### Solution
-- **Input Validation**: Added validation for matched body tag
-- **Sanitization**: Only proceed if body tag structure is valid
-- **Safe Replacement**: Extracted and validated body tag before use
-
-### Code Changes
-```javascript
-// Before (vulnerable):
-modifiedContent = modifiedContent.replace(bodyRegex, bodyMatch[0] + scriptsToAdd);
-
-// After (secure):
-const bodyTag = bodyMatch[0];
-if (bodyTag.startsWith('<body') && bodyTag.endsWith('>')) {
-  modifiedContent = modifiedContent.replace(bodyRegex, bodyTag + scriptsToAdd);
-}
-```
-
-## Expression Injection Fix
-
-### Problem
-- **Critical Vulnerability**: Expression injection in `monitor.yml` workflow
-- **Attack Vector**: Malicious branch names could execute arbitrary code
-- **Risk**: Remote code execution through workflow_run events
-- **CodeQL Alert**: "Potential injection from github.event.workflow_run.head_branch"
-
-### Solution
-- **Environment Variables**: Use `env:` block to sanitize inputs
-- **Secure Pattern**: env vars → shell vars → outputs
-- **Input Validation**: All user-controlled data properly escaped
-- **Defense in Depth**: Multiple layers of protection
-
-### Code Changes
-```yaml
-# Before (vulnerable):
-run: |
-  BRANCH="${{ github.event.workflow_run.head_branch }}"
-
-# After (secure):
-env:
-  BRANCH: ${{ github.event.workflow_run.head_branch }}
-run: |
-  echo "branch=${BRANCH}" >> $GITHUB_OUTPUT
-```
-
-### Attack Prevention
-- **Malicious Branch Names**: Names like `"; rm -rf / #"` are now safe
-- **Code Injection**: Cannot execute arbitrary commands
-- **Data Exfiltration**: Prevents information disclosure
-- **Privilege Escalation**: Blocks unauthorized access
-
-## Security Benefits
-
-### Enhanced Detection
-- **Latest Rules**: Always uses most current Semgrep rulesets
-- **Faster Scans**: Improved performance with latest version
-- **Better Integration**: Native GitHub Code Scanning integration
-
-### Vulnerability Prevention
-- **Input Validation**: Prevents injection through malformed HTML
-- **Safe Processing**: Validates all external data before use
-- **Defensive Coding**: Graceful handling of edge cases
-
-### Compliance
-- **Modern Tooling**: Uses supported, actively maintained security tools
-- **Standards Alignment**: Follows GitHub security best practices
-- **Audit Trail**: SARIF outputs provide detailed security findings
-
-## Testing
-
-- **Local Validation**: Scripts tested with various HTML structures
-- **CI Integration**: Semgrep now runs on every push and PR
-- **Error Handling**: Graceful fallbacks for all edge cases
-
-## Future Considerations
-
-- **Rule Updates**: Semgrep rules will automatically stay current
-- **Performance**: Monitor scan times with latest version
-- **Coverage**: Consider additional security tools as needed
