+++ src/__tests__/components/App.test.js
@@ -1,3 +1,4 @@
+/* eslint-disable testing-library/no-container, testing-library/no-node-access */
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import App from 'App';
