+++ src/components/Blocked/variants/index.clean.jsx
@@ -15,8 +15,9 @@ export class Blocked extends Component {
     this.state = {
       url: '', // Initialize url in state
       reason: '', // Initialize reason state
-      message: translate('defaultBlockingMessage'),
-      displayBlockedLink: true,
+      message: props.message || translate('defaultBlockingMessage'),
+      displayBlockedLink:
+        props.displayBlockedLink !== undefined ? props.displayBlockedLink : true, // Default to showing it
     };
   }
 
@@ -37,42 +38,77 @@ export class Blocked extends Component {
     if (isDevEnv && !finalUrl) {
       finalUrl = 'https://www.example.com'; // Dev fallback
     }
-    
-    const finalReason = parsedReason ? decodeURIComponent(parsedReason) : 'REASON_NOT_IN_URL_PARAMS';
+
+    const finalReason = parsedReason
+      ? decodeURIComponent(parsedReason)
+      : 'REASON_NOT_IN_URL_PARAMS'; // Default if not found
     debug.log('[Blocked Page] Initial Parsed - URL:', finalUrl, 'Reason:', finalReason);
 
-    this.setState({ 
-      url: finalUrl, 
-      reason: finalReason 
-    });
+    this.setState(
+      {
+        url: finalUrl,
+        reason: finalReason,
+      },
+      () => {
+        // Log state after URL params are set
+        debug.log(
+          '[Blocked Page] State after URL parse - URL:',
+          this.state.url,
+          'Reason:',
+          this.state.reason,
+        );
+      },
+    );
 
     if (isPageReloaded()) {
       debug.log('page reloaded!');
       if (finalUrl) {
+        // Use the parsed finalUrl
         sendMessage('isUrlStillBlocked', finalUrl).then((isUrlStillBlocked) => {
           if (isUrlStillBlocked === false) {
             sendMessage('redirectSenderTab', finalUrl);
+            // No return needed here
           }
         });
       }
     }
-    
-    // Retrieve stored settings
-    storage.get({
-      message: translate('defaultBlockingMessage'),
-      displayBlockedLink: true,
-    }).then((items) => {
-      if (items) {
-        this.setState({
-          message: items.message.length ? items.message : translate('defaultBlockingMessage'),
-          displayBlockedLink: items.displayBlockedLink !== false,
-        }, () => {
-          debug.log('[Blocked Page] State after storage.get:', this.state);
-        });
-      }
-    });
+
+    storage
+      .get({
+        message: translate('defaultBlockingMessage'),
+        displayBlockedLink: true,
+      })
+      .then((items) => {
+        if (items) {
+          this.setState(
+            {
+              message: items.message.length
+                ? items.message
+                : translate('defaultBlockingMessage'),
+              // Make sure the blocked URL is shown if enabled in settings
+              displayBlockedLink: items.displayBlockedLink,
+            },
+            () => {
+              // Log state after storage items are applied
+              debug.log(
+                '[Blocked Page] State after storage.get - URL:',
+                this.state.url,
+                'Reason:',
+                this.state.reason,
+              );
+            },
+          );
+        } else {
+          debug.log(
+            '[Blocked Page] storage.get returned no items. State remains - URL:',
+            this.state.url,
+            'Reason:',
+            this.state.reason,
+          );
+        }
+      });
   }
-  
+
   copyBlockedLink = () => {
     if (copy(this.state.url)) {
       toaster.success(translate('copiedToClipboard'), {
@@ -80,29 +116,28 @@ export class Blocked extends Component {
       });
     }
   };
-  
+
   render() {
     debug.log('[Blocked Render] Rendering block page with state:', {
       url: this.state.url,
       reason: this.state.reason,
       displayBlockedLink: this.state.displayBlockedLink,
-      message: this.state.message
+      message: this.state.message,
     });
-    
+
     // Always use a default message if none is available
     const blockMessage = this.state.message || translate('defaultBlockingMessage');
-    
+
     return (
       <Fragment>
-        {/* Always show the black page with the block message */}
+        {/* Always show the block page with the block message */}
         <div className="distract-cursor distract-select distract-overlay-container">
           <div className="distract-cursor distract-select distract-overlay">
             <div className="distract-cursor distract-select distract-info-container">
               <span className="distract-cursor distract-select distract-overlay-top-text">
                 {blockMessage}
               </span>
-              
-              {/* Show the blocked link if enabled */}
+              {/* Show the blocked link only if enabled in settings */}
               {this.state.displayBlockedLink && (
                 <span className="distract-blocked-link">
                   <input type="text" value={this.state.url || ''} readOnly />
@@ -115,18 +150,20 @@ export class Blocked extends Component {
                   </button>
                 </span>
               )}
-              
               <div className="distract-cursor distract-select distract-overlay-img"></div>
-              {/* No unblock button */}
+              {/* No unblock button, no unblock dialog */}
             </div>
           </div>
         </div>
-        
+
         {/* Development mode reason display */}
         {isDevEnv && (
           <div className="reason-container">
             <p className="text-lg text-red-500 font-bold">
-              Displaying Reason: {this.state.reason || 'No reason provided'}
+              Displaying Reason:{' '}
+              {this.state.reason
+                ? this.state.reason
+                : 'No specific reason provided in state.'}
             </p>
           </div>
         )}
@@ -134,3 +171,5 @@ export class Blocked extends Component {
     );
   }
 }
+
+export default Blocked;
