+++ src/components/Blocked/variants/index.debug.jsx
@@ -1,8 +1,8 @@
 import React, { Component, Fragment } from 'react';
-import { Pane, toaster, DuplicateIcon } from 'evergreen-ui';
+import { toaster, DuplicateIcon } from 'evergreen-ui';
 import { translate } from 'helpers/i18n';
 import { storage, sendMessage } from 'helpers/webext';
-import { debug, isDevEnv } from 'helpers/debug';
+import { isDevEnv } from 'helpers/debug';
 import { getValidUrl } from 'helpers/url';
 import { isPageReloaded } from 'helpers/block';
 import copy from 'copy-to-clipboard';
