+++ src/components/Blocked/variants/index.highcontrast.jsx
@@ -3,7 +3,6 @@ import { toaster, DuplicateIcon } from 'evergreen-ui';
 import { translate } from 'helpers/i18n';
 import { storage, sendMessage } from 'helpers/webext';
 import { isDevEnv } from 'helpers/debug';
-import { getValidUrl } from 'helpers/url';
 import { isPageReloaded } from 'helpers/block';
 import copy from 'copy-to-clipboard';
 import './styles.scss';
@@ -56,6 +55,12 @@ export class Blocked extends Component {
         console.error('[Blocked] Error decoding reason:', e);
       }
     }
+
+    // Force high-contrast dummy reason for testing if we didn't find one
+    // Comment this out for production, or leave for testing
+    // if (finalReason === 'REASON_NOT_FOUND') {
+    //   finalReason = 'TEST REASON: Pattern match *.reddit.com';
+    // }
 
     this.setState({
       url: finalUrl,
@@ -147,29 +152,33 @@ export class Blocked extends Component {
 
               <div className="distract-cursor distract-select distract-overlay-img"></div>
 
-              {/* Subtle block reason display */}
+              {/* High-contrast block reason display */}
               <div
                 style={{
                   margin: '20px auto 0',
-                  padding: '12px',
-                  backgroundColor: 'transparent',
-                  color: '#d0d0d0',
+                  padding: '15px',
+                  backgroundColor: 'rgba(255, 255, 255, 0.15)',
+                  border: '2px solid #ff9800',
+                  borderRadius: '5px',
+                  boxShadow: '0 2px 10px rgba(0, 0, 0, 0.3)',
+                  color: 'white',
                   fontFamily: 'OpenSansFont, Arial, sans-serif',
-                  fontSize: '15px',
+                  fontSize: '16px',
                   maxWidth: '80%',
                   textAlign: 'center',
                 }}
               >
                 <div
-                  style={{ marginBottom: '5px', color: '#d0d0d0', fontWeight: 'normal' }}
+                  style={{ marginBottom: '8px', color: '#ff9800', fontWeight: 'bold' }}
                 >
                   {translate('blockedDueTo')}:
                 </div>
                 <div
                   style={{
-                    color: '#db9d61',
+                    color: '#ffffff',
                     fontWeight: 'bold',
                     wordBreak: 'break-word',
+                    textShadow: '1px 1px 2px rgba(0,0,0,0.7)',
                   }}
                 >
                   {this.state.reason &&
