+++ src/components/Blocked/variants/index.simplified.jsx
